Remove unused code, consolidate duplicate input components, refresh CLAUDE.md
Drops dead code that knip flagged: unused fetch helpers (fetchTest,
fetchDbTest, fetchSingleTaskType), unused dailyHelpers exports, the
test/dbTest/hello-world API routes plus their shared Test/DbTest/DynamicTest
types, and stale re-exports from the dailies and formFields barrels.

Consolidates the duplicated components/forms/{input,textarea,input-group}.tsx
into the existing components/{input,textarea,input-group}.tsx and rewrites
all imports.
