+++ packages/client/src/components/dailies/index.ts
@@ -2,18 +2,15 @@ export { DailiesActiveListView } from "./DailiesActiveListView";
 export { DailyCriteriaTemplateEditModal } from "./DailyCriteriaTemplateEditModal";
 export { DailiesViewModeToggle } from "./DailiesViewModeToggle";
 export { DailyCommentPopover } from "./DailyCommentPopover";
-export { DailyStatusModal } from "./DailyStatusModal";
 export { DailyCompletionsManager } from "./DailyCompletionsManager";
 export { DailyCourseIndicator } from "./DailyCourseIndicator";
 export { DailyLocationCell } from "./DailyLocationCell";
 export { DailyProgressCell } from "./DailyProgressCell";
 export { DailyTaskIndicator } from "./DailyTaskIndicator";
 export { DailyRecentDaysStrip } from "./DailyRecentDaysStrip";
-export { DailyStatusButtons } from "./DailyStatusButtons";
 export { DailyStatusCircle } from "./DailyStatusCircle";
 export { DailyStatusConnector } from "./DailyStatusConnector";
-export { NoteEditButton } from "./NoteEditButton";
-export { DAILY_STATUS_OPTIONS, getDailyStatusOption } from "./dailyStatusMeta";
+export { DAILY_STATUS_OPTIONS } from "./dailyStatusMeta";
 export { TodayStatusCell } from "./TodayStatusCell";
 export { TooManyDailiesWarning } from "./TooManyDailiesWarning";
 export { DailiesLimitSetting } from "./DailiesLimitSetting";
