--- packages/client/src/components/forms/input.stories.tsx
@@ -1,15 +1,0 @@
-import type { Meta, StoryObj } from "@storybook/react-vite";
-
-import { Input } from "./input";
-
-const meta = {
-  component: Input,
-} satisfies Meta<typeof Input>;
-
-export default meta;
-
-type Story = StoryObj<typeof meta>;
-
-export const Default: Story = {
-  args: {},
-};
