+++ packages/client/src/components/radar/BlipLlmAssist.tsx
@@ -13,8 +13,8 @@ import { useMemo, useState } from "react";
 import { CheckIcon, CopyIcon, Loader2, PencilIcon, XIcon } from "lucide-react";
 import { toast } from "sonner";
 
-import { Input } from "@/components/forms/input";
-import { Textarea } from "@/components/forms/textarea";
+import { Input } from "@/components/input";
+import { Textarea } from "@/components/textarea";
 import { Button } from "@/components/ui/button";
 import {
   Select,
