+++ packages/client/src/utils/fieldContext.ts
@@ -1,6 +1,6 @@
 import { createFormHookContexts } from "@tanstack/react-form";
 
 export const {
-  fieldContext, formContext, useFieldContext, useFormContext,
+  fieldContext, formContext, useFieldContext,
 }
   = createFormHookContexts();
