--- packages/middleware/src/routes/api/dbTest.ts
@@ -1,17 +1,0 @@
-import { JsonSchemaToTsProvider } from "@fastify/type-provider-json-schema-to-ts";
-import { FastifyInstance } from "fastify";
-import { usersTable } from "@/db/schema";
-import { db } from "@/db";
-
-export default async function (server: FastifyInstance) {
-  const fastify = server.withTypeProvider<JsonSchemaToTsProvider>();
-
-  fastify.get(
-    "/dbTest",
-    async (request, reply) => {
-      const result = await db.select().from(usersTable);
-
-      return result;
-    },
-  );
-}
