+++ packages/middleware/src/routes/api/routes.ts
@@ -1,11 +1,8 @@
 import { JsonSchemaToTsProvider } from "@fastify/type-provider-json-schema-to-ts";
 import { FastifyInstance } from "fastify";
 
-import apiRoot from "./root";
-import apiTest from "./testRoute";
 import apiSeed from "./seed";
 import apiClear from "./clearData";
-import dbTest from "./dbTest";
 import apiFormSubmit from "./submitOnboardData";
 import courses from "./courses/routes";
 import topics from "./topics/routes";
@@ -20,12 +17,9 @@ import dailyCriteriaTemplates from "./daily-criteria-templates/routes";
 export default async function (server: FastifyInstance) {
   const fastify = server.withTypeProvider<JsonSchemaToTsProvider>();
 
-  fastify.register(apiRoot);
-  fastify.register(apiTest);
   fastify.register(apiSeed);
   fastify.register(apiClear);
   fastify.register(apiFormSubmit);
-  fastify.register(dbTest);
   fastify.register(courses, {
     prefix: "/courses",
   });
