--- packages/types/src/DbTest.ts
@@ -1,6 +1,0 @@
-export interface DbTest {
-  id: string;
-  name: string;
-  age: number;
-  email: string;
-}
