--- packages/types/src/DynamicTest.ts
@@ -1,1 +1,0 @@
-export type DynamicTest = string;
