--- packages/types/src/Test.ts
@@ -1,4 +1,0 @@
-export interface Test {
-  item: string;
-  fromEnv: string;
-}
