+++ packages/types/src/index.ts
@@ -1,8 +1,5 @@
 /* eslint-disable import/max-dependencies */
 export * from "./CostData";
-export * from "./DynamicTest";
-export * from "./Test";
-export * from "./DbTest";
 export * from "./Course";
 export * from "./CourseInCourses";
 export * from "./CourseFromServer";
